+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Unreleased
 ----------
 - Added support for Breakpad format behind `breakpad` feature (disabled
   by default)
+- Added support for usage of perf map files as part of process symbolization
+  - Added `perf_map` attribute to `symbolize::Process` type
 
 
 0.2.0-alpha.10