+++ capi/CHANGELOG.md
@@ -1,3 +1,4 @@
 Unreleased
 ----------
 - Renamed various symbolization functions to closer reflect Rust terminology
+- Added `perf_map` member to `blaze_symbolize_src_process` function